BUGFIX: Fixes a race condition when creating phaontom modules
Phantom modules use random numbers in their names to avoid being guessed.
However, it could happen that two processes run one just after the other end up
with the same random seed (e.g., the resolution of the time-of-day counter is
not enough) so we can get a clash of names.

We fix this by including the process id in the module name.

The source of the race-condition was discovered by:
Samuel Gélineau (gelisam@gmail.com) and Mario Pastorelli (pastorelli.mario@gmail.com)